Add files via upload
Eclipse has to be used to ensure the proper functioning as well as connectivity of the database with the main java code. SQL database is not added here as it can be created using MYSQL and the url,username and password can be set as required in the file DBConnect.